backend/requirements.txt
# Roomify backend requirements
fastapi
uvicorn[standard]
python-multipart
numpy
opencv-python
pillow
torch
torchvision
transformers
timm
colorthief
midas
segment-anything
yolov5
detectron2
scikit-image
matplotlib
open3d

# Add other model-specific dependencies as needed
